Refactor APIs.
Both firestore and spotify API has been refactored.

In the spotify API we do not use runCatching anymore. None of thd Kovibes APIs
throw exception. So if an exception does occur it needs to be checked and
fixed in the library.

Now the FirestoreApi always returns FirestoreApiResponse. So, if there is
any exception occurs when calling firestore then it is handled in the same
service.